renamed banana_memory to memory
