+++ CMakeLists.txt
@@ -12,6 +12,6 @@ add_executable(${PROJECT_NAME}
     src/main.cpp
     src/console.cpp
     src/cpu.cpp
-    src/banana_memory.cpp
+    src/memory.cpp
     src/os.cpp
 )
