+++ src/memory.cpp
@@ -1,5 +1,4 @@
-#include "banana_memory.h"
-
+#include "memory.h"
 #include "console.h"
 
 #include <cstdio>
